celery/requirements.txt
celery
pika
Pillow
jsonpath-rw
dpath
numpy
opencv-python
requests
